+++ src/server/js/publish.js
@@ -9,11 +9,49 @@ import { AccountsHub } from 'meteor/pwix:accounts-hub';
 // returns a cursor of all accounts in the collection
 Meteor.publish( 'pwix_accounts_manager_accounts_list_all', async function( instanceName ){
     const amInstance = AccountsHub.instances[instanceName];
+    const self = this;
+
+    // @param {Object} item the Record item
+    // @returns {Object} item the transformed item
+    const f_transform = async function( item ){
+        item.DYN = {};
+        const fn = amInstance.serverAllExtend();
+        if( fn ){
+            await fn( item, self.userId );
+        }
+        return item;
+    };
+
     if( amInstance && amInstance instanceof AccountsManager.amClass ){
-        if( !await AccountsManager.isAllowed( 'pwix.accounts_manager.feat.list', this.userId, { amInstance: amInstance } )){
+        if( !await AccountsManager.isAllowed( 'pwix.accounts_manager.feat.list', self.userId, { amInstance: amInstance } )){
             return false;
         }
-        return amInstance.collection().find({}, { transform: AccountsHub.cleanupUserDocument });
+        let initializing = true;
+
+        const observer = amInstance.collection().find().observeAsync({
+            added: async function( item ){
+                const transformed = await f_transform( item );
+                self.added( amInstance.collectionName(), item._id, transformed );
+            },
+            changed: async function( newItem, oldItem ){
+                if( !initializing ){
+                    const transformed = await f_transform( newItem );
+                    self.changed( amInstance.collectionName(), newItem._id, transformed );
+                }
+            },
+            removed: async function( oldItem ){
+                self.removed( amInstance.collectionName(), oldItem._id );
+            }
+        });
+
+        initializing = false;
+
+        self.onStop( function(){
+            observer.then(( handle ) => { handle.stop(); });
+        });
+
+        self.ready();
+
     } else {
         console.warn( 'pwix_accounts_manager_accounts_list_all unknown or invalid instance name', instanceName );
         return false;
@@ -76,7 +114,7 @@ Meteor.publish( 'pwix_accounts_manager_accounts_tabular', async function( tableN
             item.DYN = {};
             const fn = amInstance.serverTabularExtend();
             if( fn ){
-                await fn( item );
+                await fn( item, self.userId );
             }
             return item;
         };